refactored a little
